+++ tfg_project/app/Http/Controllers/ApiControllers/PostController.php
@@ -6,8 +6,8 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\CreatePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use App\Http\Resources\PostResource;
+use App\Models\Category;
 use App\Models\Post;
-use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
@@ -25,6 +25,17 @@ class PostController extends Controller
     }
 
     /*
+    * Devuelve los posts de los usuarios seguidos
+    */
+
+    /*public function getPostsFromFollowed(Request $request)
+    {
+        $posts = $request->user()->followed()->get();
+
+        dd($posts);
+    }*/
+
+    /*
     * Crea un post
     */
 
@@ -36,6 +47,9 @@ class PostController extends Controller
         ]);
 
         if ($post->save()) {
+
+            $post->categories()->attach($request->categorias);
+
             if ($request->image) {
                 $path = 'posts/image_' . $post->id . '.' . $request->image->getClientOriginalExtension();
                 Storage::put($path, file_get_contents($request->image));
@@ -58,31 +72,41 @@ class PostController extends Controller
     public function update(UpdatePostRequest $request, $id)
     {
         $currentUser = $request->user();
-        $post = $currentUser->posts()->findOrFail($id);
+        $post = $currentUser->posts()->find($id);
 
-        $post->fill([
-            'body' => $request->body
-        ]);
+        if ($post){
+            $post->fill([
+                'body' => $request->body
+            ]);
 
-        if ($post->update()) {
-            if ($request->image) {
-                $path = 'posts/image_' . $post->id . '.' . $request->image->getClientOriginalExtension();
+            if ($post->update()) {
 
-                if (Storage::exists($path)) {
-                    Storage::delete($path);
+                $post->categories()->sync($request->categorias);
+
+                if ($request->image) {
+                    $path = 'posts/image_' . $post->id . '.' . $request->image->getClientOriginalExtension();
+
+                    if (Storage::exists($path)) {
+                        Storage::delete($path);
+                    }
+
+                    Storage::put($path, file_get_contents($request->image));
+                    $post->update(['image' => $path]);
                 }
-
-                Storage::put($path, file_get_contents($request->image));
-                $post->update(['image' => $path]);
+                return response()->json([
+                    'message' => 'Post actualizado correctamente'
+                ], 200);
             }
+
             return response()->json([
-                'message' => 'Post actualizado correctamente'
-            ], 200);
+                'message' => 'El post no ha sido actualizado'
+            ], 400);
+        }else{
+            return response()->json([
+                'message' => 'Post no encontrado'
+            ], 400);
         }
 
-        return response()->json([
-            'message' => 'El post no ha sido actualizado'
-        ], 400);
     }
 
     /*
@@ -94,14 +118,27 @@ class PostController extends Controller
         $currentUser = $request->user();
         $post = $currentUser->posts()->findOrFail($id);
 
-        if ($post->forceDelete()) {
+        if ($currentUser){
+            if ($post){
+                Storage::delete($post->image);
+                if ($post->forceDelete()) {
+                    return response()->json([
+                        'message' => 'Eliminado correctamente'
+                    ], 200);
+                }
+                return response()->json([
+                    'message' => 'No se ha podido eliminar'
+                ], 400);
+            }else{
+                return response()->json([
+                    'message' => 'Post no encontrado'
+                ], 400);
+            }
+        }else{
             return response()->json([
-                'message' => 'Eliminado correctamente'
-            ], 200);
+                'message' => 'Usuario no encontrado'
+            ], 400);
         }
-        return response()->json([
-            'message' => 'No se ha podido eliminar'
-        ], 400);
     }
 
     /*
@@ -113,28 +150,30 @@ class PostController extends Controller
         $currentUser = $request->user();
         $post = $currentUser->posts()->findOrFail($id);
 
-        if ($post->likes()->where('user_id', $currentUser->id)->first()){
-            $post->likes()->detach($currentUser->id);
+        if ($currentUser){
+            if ($post) {
+                if ($post->likes()->where('user_id', $currentUser->id)->first()){
+                    $post->likes()->detach($currentUser->id);
+                    return response()->json([
+                        'message' => 'Dislike a post'
+                    ], 200);
+                }
+
+                $post->likes()->attach($currentUser->id);
+
+                return response()->json([
+                    'message' => 'Like a post'
+                ], 200);
+            }else{
+                return response()->json([
+                    'message' => 'Post no encontrado'
+                ], 400);
+            }
+        }else{
             return response()->json([
-                'message' => 'Dislike a post'
-            ], 200);
+                'message' => 'Usuario no encontrado'
+            ], 400);
         }
-
-        $post->likes()->attach($currentUser->id);
-
-        return response()->json([
-            'message' => 'Like a post'
-        ], 200);
     }
 
-
-    /*
-    * Follow/Unfollow a un user
-    */
-
-    public function followUnfollow(Request $request, $id)
-    {
-        $currentUser = $request->user();
-        $user = User::where()
-    }
 }
